[석훈] API 클래스 매서드로 정리
